Fixed cut plane collisions and cut plane coloring in respect to that.
Start "Cut By line" implementation (needs a help from ET)

+++ src/slic3r-shared/src/Slic3r/App/Scene/Clipper.cpp
@@ -227,7 +227,7 @@ bool Clipper::unproject_on_cut_plane(
     Vec3d point     = ray.origin;
     Vec3d direction = ray.direction;
     Vec3d hit;
-    Vec3d normal = -m_clp->get_normal().cast<double>();
+    Vec3d normal = -m_clp->get_normal();
     double den   = normal.dot(direction);
     if (den != 0.) {
         double t = (-m_clp->get_offset() - normal.dot(point)) / den;
